fix: 오늘 소비현황, 캘린더 응답의 하루 합계를 Long으로 전환

+++ src/main/java/Hampouch/server/domain/challenge/dto/CalendarResponse.java
@@ -18,15 +18,11 @@ public record CalendarResponse(
         int month,
         List<DayView> days       // 그 달 안에서 기록 있는 날만
 ) {
-    /**
-     * days 배열의 원소 = 하루치 기록. 캘린더 응답에서만 쓰는 전용 부품이라 중첩으로 소속을 명시(여러 곳에서 쓰면 별도 파일로).
-     * View 접미사 = ChallengeDay 엔티티(저장 원본)를 화면에 필요한 필드만 추려 "바라본 모습"이라는 뜻
-     * (ChallengeView와 같은 관례 — 원본이 아니라 조회 전용 투영임을 이름으로 표시).
-     */
+    //days 배열의 원소 = 하루치 기록.
     public record DayView(
             LocalDate date,
             DayStatus status,
-            int spentAmount
+            long spentAmount
     ) {
     }
 }
